Move Hydrawise constants into const.py
